Added sqlite and apsw version check.
--HG--
extra : convert_revision : svn%3A0275539e-8e52-0410-9d9c-41c7838199bd/trunk%40196

+++ src/s3ql/__init__.py
@@ -10,12 +10,25 @@ if sys.version_info[0] > 2:
     sys.stderr.write('Python version too new, must be between 2.6.0 and 3.0!\n')
     sys.exit(1)
 
-# TODO: Enforce apsw and sqlite version    
+import apsw
+tmp = apsw.apswversion()
+tmp = tmp[:tmp.index('-')]
+apsw_ver = tuple([ int(x) for x in tmp.split('.') ])
+if apsw_ver < (3, 6, 14):    
+    sys.stderr.write('APSW version too old, must be 3.6.14 or newer!\n')
+    sys.exit(1)
     
+    
+sqlite_ver = tuple([ int(x) for x in apsw.sqlitelibversion().split('.') ])
+# TODO: Change to 3.6.17 as soon as it's released
+if sqlite_ver < (3, 6, 16):    
+    sys.stderr.write('SQLite version too old, must be 3.6.17 or newer!\n')
+    sys.exit(1)
+
 
 # Python boto uses several deprecated modules, deactivate warnings for them
 import warnings
 warnings.filterwarnings("ignore", "", DeprecationWarning, "boto")
     
-__all__ = [ "common", "fs", "fsck", 'fuse', 'isodate'
+__all__ = [ "common", "fs", "fsck", 'fuse', 'isodate', 'database'
             "mkfs", 'multi_lock', 'ordered_dict', "s3", 's3cache' ]
